ReGps/ml/requirements.txt
# ===================================
# ReGPS - Machine Learning Module
# Dependencias Python
# ===================================

# Core Machine Learning
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Procesamiento Geoespacial
geopy==2.3.0
shapely==2.0.1

# API y Conexión con Laravel
fastapi==0.103.1
uvicorn[standard]==0.23.2
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.3.0

# Análisis de Series Temporales
statsmodels==0.14.0

# Visualización (para notebooks)
matplotlib==3.7.2
seaborn==0.12.2
folium==0.14.0

# Entorno de Desarrollo
jupyter==1.0.0
jupyterlab==4.0.5
ipykernel==6.25.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilidades
python-dateutil==2.8.2
tqdm==4.66.1
